:recycle: split processing
